enable de pino não mais bloqueante

+++ main/include/Utils.h
@@ -8,7 +8,7 @@ typedef struct {
    gpio_num_t pin; 		// GPIO of the pin
    float duty;			// duty cycle of the GPIO
    float freq;			// toggle frequency
-   int lifetime;
+   long long int lifetime;
 } thing;
 
 
@@ -18,5 +18,5 @@ void thingEnable(void* pvParameters);
 
 
 // receive a GPIO pin, a duty cycle and frquency to enable that pin
-void enable(gpio_num_t pino, float duty, float freq, int lifetime);
+void enable(gpio_num_t pino, float duty, float freq, long long int lifetime);
